syntax: fix empty char class bug in HIR printer
When a character class is empty, the HIR printer would emit '[]', which
is not a valid regex. (Since if a ']' immediately follows an opening
'[', then the ']' is interpreted literally and not a closing bracket.)

Instead, we write '[a&&b]'. We could also do things like '(?u:\P{any})'
or '(?-u:[\x00-\xFF])', but '[a&&b]' doesn't require any flags and also
seems really obvious: the intersection of two distinct characters is
obviously empty.